pyproject.toml
[project]
name = "codegraph-p1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.7",
    "neo4j>=6.1.0",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "tiktoken>=0.12.0",
]
